+++ app_android/gradle/libs.versions.toml
@@ -1,10 +1,10 @@
 [versions]
-agp = "8.3.2"
+agp = "8.3.1"
 junit = "4.13.2"
 junitVersion = "1.1.5"
 espressoCore = "3.5.1"
 appcompat = "1.6.1"
-material = "1.11.0"
+material = "1.10.0"
 constraintlayout = "2.1.4"
 
 [libraries]
