+++ app_android/settings.gradle.kts
@@ -21,4 +21,3 @@ dependencyResolutionManagement {
 
 rootProject.name = "TitleCase"
 include(":app")
-
