+++ src/app_android/.idea/.name
@@ -1,0 +1,1 @@
+TitleCase
