+++ src/app_android/app/src/main/java/com/example/titlecase/MainActivity.java
@@ -4,11 +4,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 import android.widget.TextView;
-import android.view.View;
-
-import android.content.ClipData;
-import android.content.ClipboardManager;
-import android.content.Context;
 
 import com.example.titlecase.databinding.ActivityMainBinding;
 
@@ -21,7 +16,6 @@ public class MainActivity extends AppCompatActivity {
 
     private ActivityMainBinding binding;
 
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -29,29 +23,14 @@ public class MainActivity extends AppCompatActivity {
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
-        binding.buttonConvert.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                // Update the TextView with the string from JNI
-                String text = binding.editText.getText().toString();
-
-
-                binding.editText.setText(titleCase(text));
-            }
-        });
-
-        binding.copyButton.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                ClipboardManager clipboard = (ClipboardManager) MainActivity.this.getSystemService(Context.CLIPBOARD_SERVICE);
-
-                String text = binding.editText.getText().toString();
-                ClipData clip = ClipData.newPlainText("label", text);
-
-                clipboard.setPrimaryClip(clip);
-            }
-        });
+        // Example of a call to a native method
+        TextView tv = binding.sampleText;
+        tv.setText(stringFromJNI());
     }
 
-    public native String titleCase(String input);
+    /**
+     * A native method that is implemented by the 'titlecase' native library,
+     * which is packaged with this application.
+     */
+    public native String stringFromJNI();
 }
